brandgenie-pro-mobile/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
httpx==0.27.0
bcrypt==4.3.0 
passlib==1.7.4
ecdsa==0.19.1 
pyasn1==0.4.8 
python-jose==3.4.0 
rsa==4.9.1 
six==1.17.0
psycopg2-binary==2.9.9
dnspython==2.7.0 
email-validator==2.2.0
uuid==1.30

